Bring ox-epub up to speed
Enable exporting via the usual way of doing it: exporting a single org
file to a single epub file. Very much in the same vein as ox-odt.

Add options to the derived backends. Add a menu entry for ex-epub in
org-modes export popup.

Add a dynamic variable org-epub-zip-dir to hold the current export
directory.

Remove headline caching code from org-epub-template and add in code to
read options from the org file and to generate both content.opf and
toc.ncx.

org-epub-export-to-epub: new function. Main entry point to the export
process.

org-epub-zip-it-up2: new function. Scheduled to replace
org-epub-zip-it-up.

org-epub-generate-toc-single: new function, generate the toc for a
single file epub.

This work is to address #4 and #5.

This relates to #3, since we bump Emacs dependency to 24.3.