Add style capabilities
This commit add a default style (org-epub-style-default) derived from
org-html-style-default and always adds this style to the epub.

When you define the EPUBSTYLE option in your org file, this option
will overwrite the default style and allow you to include specific
stylesheets to your liking.

We use our own style, instead of the default style from ox-html,
because the EPUB specification actually subsets CSS and doesn't for
example allow absolute positioning in the included CSS, which is
present in the ox-html default style.

In response to #7