Add manifest data type
This add a defvar `org-epub-manifest' that represents the manifest.

`org-epub-manifest' is a list of plists, each plist contains at least
the fields :filename, :type, and :mimetype and possibly :source.

Use `org-epub-manifest-entry' to create new entries.

This also adds a set of predicates to test for the type of the
manifest entry.

Additionally there are two functions to access the manifest, each
accepting a predicate, they're `org-epub-manifest-first' to get the
first entry for which the predicate is true and
`org-epub-manifest-all' to get all entries for which the predicate is
true.